fix folder style and mysql driver update
