+++ .idea/MysqlEconomyBridge.iml
@@ -13,7 +13,10 @@
     <output url="file://$MODULE_DIR$/target/classes" />
     <output-test url="file://$MODULE_DIR$/target/test-classes" />
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/MySQL Eco Bridge 3/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/MySQL Eco Bridge 3/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/MySQL Eco Bridge 3/src/main/resources" type="java-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/resources" type="java-resource" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
     </content>
     <orderEntry type="inheritedJdk" />
