--- MySQL Eco Bridge 3/bin/.gitignore
@@ -1,1 +1,0 @@
-/net/
