--- MySQL Eco Bridge 3/src/resources/config.yml
@@ -1,39 +1,0 @@
-# default config.yml for MysqlEcoBridge
-# config version 3.4.1
-
-#MySQL Database details
-database:
-    mysql:
-        #MySQL server address
-        host: 127.0.0.1
-        #MySQL server port (default 3306)
-        port: 3306
-        #Database name (NOTE! You need to create the database, then the plugin will create the tables.)
-        databaseName: 'mydatabase'
-        #Tables name (the plugin will auto create it)
-        dataTableName: 'eco_accounts'
-        #User name
-        user: 'admin'
-        #User password
-        password: 'cheesecake'
-        #SSL connection.
-        sslEnabled: false
-    #This maintenance task runs async with a 2 min delay after the server starts.
-    removeOldAccounts:
-        #Enable or disable database clean up of old accounts. | (true or false)
-        enabled: false
-        #Inactivity in days. Default 60 days.
-        inactivity: 60
- 
- 
-#Other configurable options
-General:
-    #Save online players accounts to database task. Disable this task and data will only be saved when the player disconnects.
-    #To limit data loss in case the server crashes enable this saving task. It runs async so there will be no lag involved.
-    saveDataTask:
-        #Enable or disable the data save task. | (true or false)
-        enabled: true
-        #Time between data saves in minutes. | Default 3 min.
-        interval: 3
-        #Hide the data save task log messages.
-        hideLogMessages: false
