+++ pom.xml
@@ -3,9 +3,8 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>net.craftersland.eco</groupId>
-    <artifactId>MysqlEconomyBridge</artifactId>
+    <groupId>net.craftersland.ecobridge</groupId>
+    <artifactId>MysqlEcoBridge</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
